+++ noodles-tabix/src/async/io/writer/index/header/reference_sequence_names.rs
@@ -1,5 +1,8 @@
+use bstr::{BStr, ByteSlice};
 use noodles_csi::binning_index::index::header::ReferenceSequenceNames;
 use tokio::io::{self, AsyncWrite, AsyncWriteExt};
+
+const NUL: u8 = 0x00;
 
 pub(super) async fn write_reference_sequence_names<W>(
     writer: &mut W,
@@ -8,19 +11,28 @@ pub(super) async fn write_reference_sequence_names<W>(
 where
     W: AsyncWrite + Unpin,
 {
-    const NUL: u8 = b'\x00';
-
     // Add 1 for each trailing NUL.
     let len: usize = reference_sequence_names.iter().map(|n| n.len() + 1).sum();
     let l_nm = i32::try_from(len).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
     writer.write_i32_le(l_nm).await?;
 
     for reference_sequence_name in reference_sequence_names {
+        if !is_valid(reference_sequence_name.as_ref()) {
+            return Err(io::Error::new(
+                io::ErrorKind::InvalidInput,
+                "invalid reference sequence name",
+            ));
+        }
+
         writer.write_all(reference_sequence_name).await?;
         writer.write_u8(NUL).await?;
     }
 
     Ok(())
+}
+
+fn is_valid(s: &BStr) -> bool {
+    s.find_byte(NUL).is_none()
 }
 
 #[cfg(test)]
@@ -62,6 +74,13 @@ mod tests {
                 b's', b'q', b'1', 0x00, // names[1] = "sq1"
             ]
         );
+
+        let reference_sequence_names = [BString::from("sq\x000")].into_iter().collect();
+        buf.clear();
+        assert!(matches!(
+            write_reference_sequence_names(&mut buf, &reference_sequence_names).await,
+            Err(e) if e.kind() == io::ErrorKind::InvalidInput
+        ));
 
         Ok(())
     }